backend/pyproject.toml
[project]
name = "docflow-ai"
version = "0.1.0"
description = "LLM + Agent 기반 올인원 문서 자동화 서비스"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.14.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    "structlog>=24.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "supabase>=2.5.0,<2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
